Desktop/virtuals_Environment/Platform_Diaspora_Ivoirienne/src/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
Pillow==10.3.0
channels==4.1.0
channels-redis==4.2.0
redis==5.0.6
django-filter==24.2
drf-spectacular==0.29.0
gunicorn==22.0.0
whitenoise==6.7.0
